backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.0
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.7
Deprecated==1.2.14
durationpy==0.9
faiss-cpu==1.9.0
fastapi==0.109.0
filelock==3.16.1
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.10.0
google-ai-generativelanguage==0.4.0
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.3.2
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.67.1
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.4.5
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kubernetes==31.0.0
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.9
langchain-google-genai==0.0.5
langchain-text-splitters==0.3.2
langsmith==0.0.92
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.3
oauthlib==3.2.2
onnxruntime==1.20.0
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.11
overrides==7.7.0
packaging==23.2
pillow==11.0.0
posthog==3.7.0
propcache==0.2.0
proto-plus==1.25.0
protobuf==4.25.5
pulsar-client==3.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.5.3
pydantic-settings==2.6.1
pydantic_core==2.14.6
Pygments==2.18.0
pyparsing==3.2.0
pypdf==3.17.4
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==2.2.2
sentencepiece==0.2.0
setuptools==75.3.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.35.1
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.5.1
torchvision==0.20.1
tqdm==4.66.6
transformers==4.46.2
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.25.0
watchfiles==0.24.0
websocket-client==1.8.0
websockets==13.1
wrapt==1.16.0
yarl==1.17.1
zipp==3.20.2
